[fix][feature]차단바 이벤트 재생 시 CarEventCanas 출력되는 문제 및 BoomGatePlayer의 시작 위치, 충돌판정 수정(BoomGatePlayer의 태그 변경), 차단바와 적정거리 도달시 속도0 입력불가
