--- Assets/BoomGateEventStart.signal.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e91ac0b76de956442ba5f678f27951d1
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
