--- Assets/Scripts/BoomGateEvent/BoomGateEventTimeline.playable.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 836b9c8164df3f542b1b757443d82fdf
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
