+++ Assets/Scripts/BoomGateEvent/BoomGateEventTrigger.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: e40d8ee35d2a33c4cb422deda95a9fab
