--- Assets/Scripts/BoomGateEvent/BoomGateTimelineController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 2ef7f966643da3047b4357089b846e9d
